[GH-90] move utility functions into separate namespace
