+++ test/util/test_wrapper.py
@@ -4,7 +4,6 @@ from impyute.util import preprocess
 from impyute.imputation.cs import mean
 from impyute.util import BadInputError
 from impyute.util import checks
-from impyute.util.wrapper import thread
 
 # pylint:disable=redefined-builtin
 try:
@@ -121,18 +120,3 @@ def test_nan_exists():
     with pytest.raises(BadInputError) as excinfo:
         some_fn(np.array([[1.]]))
     assert str(excinfo.value) == "No NaN's in given data"
-
-def _add_one(x):
-    """ """
-    return x + 1
-
-def _square(x):
-    return x * x
-
-def test_thread():
-    assert 10 == thread(3, _square, _add_one)
-    assert 100 == thread(3, _square, _add_one, _square) #4
-    assert 82 == thread(3, _square, _square, _add_one) #4
-    assert 10 == thread(3, lambda x: x*x, lambda x: x+1)
-    assert 100 == thread(3, lambda x: x*x, lambda x: x+1, lambda x: x*x)
-    assert 82 == thread(3, lambda x: x*x, lambda x: x*x, lambda x: x+1)
